alinux: iocost: fix format mismatch build warning
This fixes the following format build warning:
  block/blk-iocost.c: In function 'ioc_stat_prfill':
  block/blk-iocost.c:2506:17: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 9 has type 'long int' [-Wformat=]

Reported-by: kbuild test robot <lkp@intel.com>
Fixes: 5494d4e748cf ("alinux: iocost: add ioc_gq stat")
Acked-by: Caspar Zhang <caspar@linux.alibaba.com>

+++ block/blk-iocost.c
@@ -2508,7 +2508,7 @@ static u64 ioc_stat_prfill(struct seq_file *sf, struct blkg_policy_data *pd,
 		   dname, !list_empty(&iocg->active_list),
 		   iocg->active, iocg->inuse,
 		   iocg->hweight_active, iocg->hweight_inuse,
-		   atomic64_read(&ioc->vtime_rate));
+		   (unsigned long long)atomic64_read(&ioc->vtime_rate));
 
 	return 0;
 }
